Initial release
This file has had the testing lines stripped and modifications made to the input argument parsing. Currently, this works in a way that is satisfactory. However, I have done some more testing and found a 97.8% identity scenario where, in reality, we should accept the indel. Thus, I think I will adopt the consensus feature I used for the gene rescue module which will allow us to accept indels with reduced identity % (e.g., if only one transcript alignment we will enforce 98%, but if there are multiple that all agree on the indel location, we can accept a lower % (somewhere from 95-97% ?)