Code tidying and BLAT parsing
At this stage I believe the program is fully functional. It is a bit tidier than it was (could probably use some more...) and is now capable of parsing gene names from BLAT gff3 files. GMAP should still be the program that is actually used, however, since the two programs don't produce identical results when run through this program.