Still modifying code behaviour
There's a lot of little oddities littered throughout the code's behaviour. Currently I am attempting to refine this behaviour to handle a test contig and have trialed out various modifications. As of now I believe most of the behaviour is reasonably tuned, but there are still a few things that need to be tested.