Main module final behaviour tuning start
Since all the major features of the main module exist, all that remains is to add checks into the system to remove fringe cases that are likely to be false alterations. The first check I've added was to handle a scenario I found where a change was made that was not likely to be "false" in as much as it simply altered a part that wasn't likely to be part of the real exon.