Ready for use
I think I've finished with all the testing. I'm sure there's a few minor (and extremely rare) scenarios that I would prefer the program handle in a different way, but I've tested it extensively and believe that the program's behaviour is robust enough to handle things in a way such that the end result will be significantly improved upon the original genome sequence. I've stripped all the testing bits out of the code and it is now ready for use.