Gene and exon rescue modules
Changed the previous "gene rescue" module to be the "exon rescue" module, and added in an actual "gene rescue" module. This works on the basis of parsing a GMAP gff3 file and pulling out transcripts which, in all likelihood, should correspond to gene models in the annotation but don't. We can insert them into the main program's behaviour to identify indel errors that may have prevented these from being annotated in the first place. Parsing the GMAP file is a bit time consuming, but I think it offers a real opportunity to fix gene models that may otherwise be overlooked by the program.