Transitioning towards NCLS indexing
The newly added gene rescue module simply takes too long. It has required the program to be optimised in order to run in a reasonable time frame. To accomplish this, I am using a NCLS implementation (https://github.com/hunt-genes/ncls). This is A LOT faster, but there's currently a few small bugs I need to iron out and I need to re-do some of the functions in order to fully transition the program over to use this function.