Ready to use v2
Program fully incorporates NCLS indexing now. Program behaviour for gene rescue has been completely overhauled utilising the speed increases NCLS indexing has brought about and it works significantly better now. Specifically, it is capable of identifying genes that "should" be annotated but aren't, as well as individual exons that "should" be part of gene models but aren't. Haven't tested the speed up in the code, but the original data set finished processing in 7 hrs and it looks like this should finish in less than 1.